cogl-clip-stack: Store window rect entries as ints not floats
The window rectangles are passed in as integers so there is no point
in converting them to floats when storing a stack entry for them.

+++ cogl/cogl-clip-stack.c
@@ -129,10 +129,10 @@ struct _CoglClipStackEntryWindowRect
   CoglClipStackEntry     _parent_data;
 
   /* The window space rectangle for this clip */
-  float                  x0;
-  float                  y0;
-  float                  x1;
-  float                  y1;
+  int                    x0;
+  int                    y0;
+  int                    x1;
+  int                    y1;
 };
 
 struct _CoglClipStackEntryPath
